+++ metadata.rb
@@ -4,7 +4,7 @@ maintainer_email 'cookbooks@express42.com'
 license 'all_rights'
 description 'Installs/Configures pgbarman'
 long_description 'Installs/Configures pgbarman'
-version '0.0.2'
+version '0.0.3'
 
 depends 'cron'