+++ CHANGELOG.md
@@ -5,3 +5,7 @@
 * (new) pgbarman_bidirectional_ssh lwrp for bidirect ssh configuration
 * (new) cron scheduling for full backups
 * (fix) minor fixes and typos
+
+# 0.0.3 (18 Feb 2016)
+* (fix) config permissions
+* (fix) move default recipe to server